+++ public/service-worker.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Update cache names any time any of the cached files change.
-const CACHE_NAME = 'static-cache-v1.0.7.9';
+const CACHE_NAME = 'static-cache-v1.0.7.14';
 
 // Add list of files to cache here.
 const FILES_TO_CACHE = [