+++ src/components/TimeUp.svelte
@@ -17,14 +17,14 @@
         </div>
         <div class="row justify-content-center pt-3">
             <div class="col-12 mb-5 text-center">
-                <h5 class="c-white">Tijd is op</h5>
+                <h5 class="c-white">Tijd is op!</h5>
             </div>
         </div>
         <div class="row justify-content-center pt-3">
             <div class="col-12 col-md-8 col-lg-6 mb-3">
                 <div class="card bg-blue" on:click="{() => checkGuessedOverview()}">
                     <div class="card-body text-center">
-                        <h2 class="c-white">verder</h2>
+                        <h2 class="c-white">Verder <i class="fas fa-chevron-right"></i></h2>
                     </div>
                 </div>
             </div>