+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-## [1.0.0] - 2020-07-10
+## [1.1.0] - 2020-02-08
+### Added
+- Changelog
+- Possibility to end turn (Thanks for the idea Jip v H)
+- Help Screen
+- Wall of Fame
+- Many words
+- Possibility to delete Teams
+- Possibility to delete Players
+- Useful messages when trying to start game without enough Teams / Players
+
+### Fixed
+- Fixed Footer in settings
+- 
+
+### Changed
+- Merged App version & Words version in one
+- Force Update button
+
+
+## [1.0.0] - 2021-02-05
 ### Added
 - First Version
 
@@ -6,4 +26,4 @@
 - Nothing
 
 ### Changed
-- Refactor JavaScript wrapper
+- Nothing